EnvVars (#56)
Moved envVars and related function to a separate file.
